third_party/WebKit/LayoutTests/fast/dom/inert/inert-node-is-unselectable.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
  <div inert>Here is a text node you can't select.</div>
  <div>I'm selectable.</div>
<script>
test(function() {
    document.execCommand('SelectAll');
    assert_equals(window.getSelection().toString(), "I'm selectable.");
}, "Inert nodes cannot be selected.");
</script>
</body>
</html>
